feat(#1679): S2 — incremental BTI partition-trie emission (single sweep + depth-≤9 stack)
Replace the whole-tree-then-serialize PartitionsTrieWriter driver with a
single left-to-right sweep over the sorted, fixed-length (9-byte) keys,
holding only the root-to-current-leaf path resident on a depth-≤9 pending
stack — never an O(partitions) nested BTreeMap tree.

- New emit_partitions_trie(): LCP-driven pop/push of pending internal nodes,
  reusing the unchanged write_leaf/write_sparse/write_dense serializers, so
  output is byte-identical to the recursive post-order walk.
- The old build_trie/insert/write_node whole-tree machinery is retained only
  as a #[cfg(test)] byte-identity oracle.
- Byte-identity tests: emit == reference over single/last-byte/first-byte/
  all-same-prefix/dense-depth-8/dense-interior/max-divergent/pseudorandom/
  realistic-murmur shapes, plus a pinned full finish() golden.
- dhat peak-heap budget test (1M partitions): 128 MiB budget; post-fix peak
  ~64 MiB vs pre-fix ~1921 MiB whole-tree resident.
- Split the inline test module into partitions_writer_tests.rs (campsite
  rule, epic #1135).

Rows.db row-index trie (build_row_trie/write_row_node) is left unchanged —
out of S2 scope.

Closes #1679
